Refactor some enum into a constants file.
