+++ backend/src/modules/idim-webservice/constants.ts
@@ -1,0 +1,35 @@
+export enum SearchUserParameterType {
+    UserGuid = 'userGuid',
+    UserId = 'userId',
+}
+
+export enum RequesterAccountTypeCode {
+    Internal = 'Internal',
+    Business = 'Business',
+}
+
+export enum SearchMatchMode {
+    Exact = 'Exact',
+    Contains = 'Contains',
+    StartsWith = 'StartsWith',
+}
+
+export enum SoapMatchMode {
+  Exact = 'Exact',
+  Contains = 'Contains',
+  StartsWith = 'StartsWith',
+}
+
+export enum SoapSortDirection {
+  Ascending = 'Ascending',
+  Descending = 'Descending',
+}
+
+export enum SoapSortProperty {
+  UserId = 'UserId',
+}
+
+export enum SoapSearchResultCode {
+  Success = 'Success',
+  Failed = 'Failed',
+}
